Fix chat_id column type: INTEGER → BIGINT
Telegram chat IDs are 64-bit integers; Postgres INTEGER is 32-bit (max ~2.1B).
The INSERT into checkin_state failed with NumericValueOutOfRange on first trigger.

+++ bot/state.py
@@ -9,7 +9,7 @@ DATABASE_URL = os.environ.get("DATABASE_URL", "")
 
 SCHEMA = """
 CREATE TABLE IF NOT EXISTS checkin_state (
-    chat_id       INTEGER PRIMARY KEY,
+    chat_id       BIGINT PRIMARY KEY,
     session_idx   INTEGER NOT NULL DEFAULT 0,
     exercise_idx  INTEGER NOT NULL DEFAULT 0,
     awaiting_note INTEGER NOT NULL DEFAULT 0,
